Adapt Changes
