+++ bot/constants/globals.py
@@ -1,6 +1,5 @@
 import os
 import logging
-from enum import Enum
 
 # LOGGING
 logging.basicConfig(
@@ -60,11 +59,3 @@ JOB_INTERVAL_IN_SECONDS = 5 if DEBUG else 30
 
 # SETTINGS
 SLASH_POINTS_NOTIFICATION_THRESHOLD_DEFAULT = 3
-
-
-class NetworkHealthStatus(Enum):
-    INEFFICIENT = "Inefficient"
-    OVERBONDED = "Overbonded"
-    OPTIMAL = "Optimal"
-    UNDBERBONDED = "Underbonded"
-    INSECURE = "Insecure"
