+++ bot/constants/messages.py
@@ -1,10 +1,21 @@
-from constants.globals import HEALTH_EMOJIS, NetworkHealthStatus
+from enum import Enum
+
+from constants.globals import HEALTH_EMOJIS
 
 NETWORK_ERROR = '😱 There was an error while getting data 😱\nAn API endpoint is down!'
 HEALTH_LEGEND = f'\n*Node health*:\n{HEALTH_EMOJIS[True]} - *healthy*\n{HEALTH_EMOJIS[False]} - *unhealthy*\n' \
                 f'{HEALTH_EMOJIS[None]} - *unknown*\n'
 
-NETWORK_HEALTH_CURATION = "The network is safe and efficient again! ✅"
+
+class NetworkHealthStatus(Enum):
+    INEFFICIENT = "Inefficient"
+    OVERBONDED = "Overbonded"
+    OPTIMAL = "Optimal"
+    UNDBERBONDED = "Underbonded"
+    INSECURE = "Insecure"
+
+
+NETWORK_HEALTHY_AGAIN = "The network is safe and efficient again! ✅"
 
 
 def NETWORK_HEALTH_WARNING(network_health_status: NetworkHealthStatus) -> str:
