--- test/unit_tests/_trial_temp/test.log
@@ -1,1 +1,0 @@
-2021-01-12 09:56:14+0100 [-] Log opened.
